backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlmodel==0.0.14
alembic==1.13.1
aiosqlite==0.19.0

# Async task processing
celery==5.3.6
redis==5.0.1
flower==2.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-jwt==4.1.0

# HTTP client for metadata APIs
httpx==0.26.0

# Configuration management
pydantic-settings==2.1.0
PyYAML==6.0.1

# DVD/udev monitoring
pyudev==0.24.1

# SSH transfer
paramiko==3.4.0

# Utilities
python-magic==0.4.27
psutil==5.9.8

# For SSE (Server-Sent Events)
sse-starlette==2.0.0
